Getting to a certain level of polish
This commit kinda-sorta gives a readme that's functional, but is still lacking in almost any reasonable respect. Someday...

It also fixes some minor things with the estimators for dealing with shapes correctly/nicely.  This still has some way to go as well, but works well enough for single variate timeseries type data.

Finally, this adds some much needed stuff to analysis for the estimation of dyadic process network relations, and includes a shuffle test for elimination of spurious relations.  Currently the shuffle-test is somewhat ad-hoc, further improvements and rigor is needed.
Okay that's it!